Had to add creation of DEPLOY_DIR

+++ copy.dirs.sh
@@ -3,9 +3,14 @@
 ./gradlew dist
 
 TEST_DIR=../data/testfiles/docker-services
+DEPLOY_DIR=../docker/deploy/data/docker-services
 if [ ! -d ${TEST_DIR} ]; then
         echo "Creating docker-services dir (${TEST_DIR})"
         mkdir ${TEST_DIR} || exit 1
+fi
+if [ ! -d ${DEPLOY_DIR} ]; then
+        echo "Creating docker-services dir (${DEPLOY_DIR})"
+        mkdir ${DEPLOY_DIR} || exit 1
 fi
 
 rm -rf ${TEST_DIR}/pipelines ${TEST_DIR}/nextflow
@@ -13,7 +18,6 @@ cp -r build/dist/pipelines ${TEST_DIR}
 cp -r build/dist/nextflow ${TEST_DIR}
 echo "Files copied to ${TEST_DIR}"
 
-DEPLOY_DIR=../docker/deploy/data/docker-services
 rm -rf ${DEPLOY_DIR}/pipelines ${DEPLOY_DIR}/nextflow
 cp -r build/dist/pipelines ${DEPLOY_DIR}
 cp -r build/dist/nextflow ${DEPLOY_DIR}
